feat: move schemas to separate files
